+++ spring-cloud-starter-tencent-polaris-discovery/src/main/java/com/tencent/cloud/polaris/discovery/refresh/PolarisRefreshApplicationReadyEventListener.java
@@ -21,7 +21,7 @@ import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.TimeUnit;
 
-import com.tencent.cloud.polaris.discovery.PolarisDiscoveryHandler;
+import com.tencent.cloud.polaris.context.PolarisSDKContextManager;
 import com.tencent.polaris.client.util.NamedThreadFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -32,7 +32,6 @@ import org.springframework.cloud.client.discovery.event.HeartbeatEvent;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.context.ApplicationEventPublisherAware;
 import org.springframework.context.ApplicationListener;
-import org.springframework.lang.NonNull;
 
 import static com.tencent.cloud.polaris.discovery.refresh.PolarisServiceStatusChangeListener.INDEX;
 
@@ -46,13 +45,14 @@ public class PolarisRefreshApplicationReadyEventListener
 
 	private static final Logger LOG = LoggerFactory.getLogger(PolarisRefreshApplicationReadyEventListener.class);
 	private static final int DELAY = 60;
-	private final PolarisDiscoveryHandler polarisDiscoveryHandler;
+	private final PolarisSDKContextManager polarisSDKContextManager;
 	private final PolarisServiceStatusChangeListener polarisServiceStatusChangeListener;
 	private final ScheduledExecutorService refreshExecutor;
 	private ApplicationEventPublisher publisher;
 
-	public PolarisRefreshApplicationReadyEventListener(PolarisDiscoveryHandler polarisDiscoveryHandler, PolarisServiceStatusChangeListener polarisServiceStatusChangeListener) {
-		this.polarisDiscoveryHandler = polarisDiscoveryHandler;
+	public PolarisRefreshApplicationReadyEventListener(PolarisSDKContextManager polarisSDKContextManager,
+			PolarisServiceStatusChangeListener polarisServiceStatusChangeListener) {
+		this.polarisSDKContextManager = polarisSDKContextManager;
 		this.polarisServiceStatusChangeListener = polarisServiceStatusChangeListener;
 		this.refreshExecutor = Executors.newSingleThreadScheduledExecutor(
 				new NamedThreadFactory("polaris-service-refresh"));
@@ -61,7 +61,7 @@ public class PolarisRefreshApplicationReadyEventListener
 	@Override
 	public void onApplicationEvent(ApplicationReadyEvent event) {
 		// Register service change listener.
-		polarisDiscoveryHandler.getSdkContext().getExtensions().getLocalRegistry()
+		polarisSDKContextManager.getSDKContext().getExtensions().getLocalRegistry()
 				.registerResourceListener(polarisServiceStatusChangeListener);
 
 		// Begin scheduled refresh thread.
@@ -84,12 +84,12 @@ public class PolarisRefreshApplicationReadyEventListener
 	}
 
 	@Override
-	public void setApplicationEventPublisher(@NonNull ApplicationEventPublisher applicationEventPublisher) {
+	public void setApplicationEventPublisher(ApplicationEventPublisher applicationEventPublisher) {
 		this.publisher = applicationEventPublisher;
 	}
 
 	@Override
-	public void destroy() throws Exception {
+	public void destroy() {
 		refreshExecutor.shutdown();
 	}
 }