+++ spring-cloud-starter-tencent-polaris-router/src/test/java/com/tencent/cloud/polaris/router/config/LoadBalancerConfigurationTest.java
@@ -21,7 +21,6 @@ package com.tencent.cloud.polaris.router.config;
 import com.tencent.cloud.polaris.context.config.PolarisContextAutoConfiguration;
 import com.tencent.cloud.polaris.router.PolarisRouterServiceInstanceListSupplier;
 import com.tencent.cloud.rpc.enhancement.transformer.PolarisInstanceTransformer;
-import com.tencent.polaris.router.api.core.RouterAPI;
 import org.junit.jupiter.api.Test;
 
 import org.springframework.boot.autoconfigure.AutoConfigurations;
@@ -63,7 +62,6 @@ public class LoadBalancerConfigurationTest {
 				.withBean(PolarisInstanceTransformer.class)
 				.run(context -> {
 					assertThat(context).hasSingleBean(LoadBalancerConfiguration.PolarisReactiveSupportConfiguration.class);
-					assertThat(context).hasSingleBean(RouterAPI.class);
 					assertThat(context).hasSingleBean(ReactiveDiscoveryClient.class);
 					assertThat(context).hasSingleBean(PolarisRouterServiceInstanceListSupplier.class);
 				});
@@ -81,7 +79,6 @@ public class LoadBalancerConfigurationTest {
 				.run(context -> {
 					assertThat(context).hasSingleBean(LoadBalancerConfiguration.PolarisBlockingSupportConfiguration.class);
 					assertThat(context).hasSingleBean(DiscoveryClient.class);
-					assertThat(context).hasSingleBean(RouterAPI.class);
 					assertThat(context).hasSingleBean(PolarisRouterServiceInstanceListSupplier.class);
 				});
 	}